Rename new workshop tests
